requirements.txt
flask>=3.0.0
flask-cors>=5.0.0
openai>=1.0.0
numpy>=1.26.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
gunicorn>=23.0.0
pydub>=0.25.1
webvtt-py>=0.5.1